#26 継承を利用して、重複コードを削除、Francのequalsメソッドのテストを追加
