+++ shino/src/main/java/org/codepolaris/tdd/Dollar.java
@@ -1,9 +1,7 @@
 package org.codepolaris.tdd;
 
-import java.util.Objects;
 
-public class Dollar {
-  private int amount;
+public class Dollar extends Money {
 
   public Dollar(int amount) {
     this.amount = amount;
@@ -11,18 +9,5 @@ public class Dollar {
 
   public Dollar times(int multiplier) {
     return new Dollar(amount * multiplier);
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    Dollar dollar = (Dollar) o;
-    return amount == dollar.amount;
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(amount);
   }
 }
