+++ shino/src/main/java/org/codepolaris/tdd/Franc.java
@@ -1,9 +1,6 @@
 package org.codepolaris.tdd;
 
-import java.util.Objects;
-
-public class Franc {
-  private int amount;
+public class Franc extends Money {
 
   public Franc(int amount) {
     this.amount = amount;
@@ -11,18 +8,5 @@ public class Franc {
 
   public Franc times(int multiplier) {
     return new Franc(amount * multiplier);
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    Franc franc = (Franc) o;
-    return amount == franc.amount;
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(amount);
   }
 }
