+++ shino/src/main/java/org/codepolaris/tdd/Money.java
@@ -1,0 +1,20 @@
+package org.codepolaris.tdd;
+
+import java.util.Objects;
+
+public class Money {
+  protected int amount;
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    Money money = (Money) o;
+    return amount == money.amount;
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(amount);
+  }
+}
