+++ shino/src/test/java/org/codepolaris/tdd/MoneyTest.java
@@ -17,6 +17,8 @@ class MoneyTest {
   void testEquality() {
     assertThat(new Dollar(5).equals(new Dollar(5))).isTrue();
     assertThat(new Dollar(5).equals(new Dollar(6))).isFalse();
+    assertThat(new Franc(5).equals(new Franc(5))).isTrue();
+    assertThat(new Franc(5).equals(new Franc(6))).isFalse();
   }
 
   @Test
